Getting Student Location
The javascript now can get the student location. But the student form and location is not being saved in db. The allow location and finish buttons do not work as intended(we want the allow location button to be enabled when both checkboxes are clicked(this works), and we want the finish button to go to next frame(frame 4) when the allow location is clicked and the form is submitted to backend db successfully) Also, I do not know why the url's generated by the app for event specific have random ip address or wrong rest of the url as of this commit.